HSEARCH-2620 Move JBeret specific file
The file:

    hibernate-search-jsr352-core/src/main/resources/META-INF/services/org.jberet.spi.JobXmlResolver

should move to the `hibernate-search-jsr352-jberet` module, because it
is an implementation-specific file. This file is required by WildFly
because our JSR-352 job is installed via a module, and it not part of a
deployment. It allows this job to be used in WildFly without asking the
user to rebuild or modify an existing JAR.