HSEARCH-2615 Fix the greater-or-equal operator for CompositeIdOrder
With PX as the property names, and VX as those properties' values
in the compared ID, we used to translate the greater-or-equal operator
as P1 >= V1 OR P1 = V1 AND P2 >= V2,
but that's absurd as it's equivalent to P1 >= V1 (the first clause
will always be true if the second one is true).

The actual condition we were looking for was
P1 > V1 OR P1 = V1 AND P2 > V2 OR P1 = V1 AND P2 = V2.

This commit also slightly changes the tests in order to
reproduce the bug, which used to happen in
MassIndexingJobWithCompositeIdTest.canHandleIdClass_strategyFull()
for instance, where partitions 2 ([2017-06-14,2017-07-20[) and
3 ([2017-07-20,2017-06-27[) both processed the date 2017-06-20
(with the lexicographical order ("day", "month", "year").